Rename tests Entity -> Drawable, Added new test names to the automated perf tests.
