+++ Scripts/_AutomatedPeriodicBenchmarkSuite_.bv.lua
@@ -14,7 +14,8 @@ FRAME_COUNT = 100
 SAMPLES_TO_RUN = {
     {prefix = 'RPI', name = 'CullingAndLod', width = 1400, height = 800},
     {prefix = 'RPI', name = 'SponzaBenchmark', width = 1400, height = 800},
-    {prefix = 'RPI', name = 'HighInstanceTest', width = 800, height = 600}
+    {prefix = 'Performance', name = '100KDrawable_SingleView', width = 800, height = 600}
+    {prefix = 'Performance', name = '100KDraw_10KDrawable_MultiView', width = 800, height = 600}
 }
 
 Print('Capturing data for ' .. tostring(#SAMPLES_TO_RUN) .. ' benchmarks')
